Features 5,6

+++ src/main/java/com/dkit/gd2/dominikHampejs/DAO/MySqlChampionDAO.java
@@ -7,9 +7,26 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
 
 
 public class MySqlChampionDAO extends MySqlDAO implements IChampionDAO {
+
+    private static HashSet<Integer> ids = new HashSet<>();
+
+    public MySqlChampionDAO() {
+        super();
+        List<Champion> champions = null;
+        try {
+            champions = findAllChampions();
+        } catch (DAOexception e) {
+            System.out.println("Couldn't build the cache. " + e.getMessage());
+        }
+        for(Champion c : champions){
+            ids.add(c.getId());
+        }
+    }
 
     @Override
     public List<Champion> findAllChampions() throws DAOexception {
@@ -57,32 +74,34 @@ public class MySqlChampionDAO extends MySqlDAO implements IChampionDAO {
         ResultSet rs = null;
         Champion c = null;
 
-        try{
-            con = this.getConnection();
-            String query = "SELECT * FROM champion WHERE id = ?";
-            ps = con.prepareStatement(query);
-            ps.setInt(1, id);
-            rs = ps.executeQuery();
+        if (ids.contains(id)) {
+            try {
+                con = this.getConnection();
+                String query = "SELECT * FROM champion WHERE id = ?";
+                ps = con.prepareStatement(query);
+                ps.setInt(1, id);
+                rs = ps.executeQuery();
 
-            if(rs.next()){
-                String name = rs.getString("name");
-                String role = rs.getString("role");
-                double winRate = rs.getDouble("winRate");
-                c = new Champion(id, name, role, winRate);
-            }
-        } catch (Exception e) {
-            throw new DAOexception("findChampionById() " + e.getMessage());
-        } finally {
-            try{
-                if(rs != null){
-                    rs.close();
+                if (rs.next()) {
+                    String name = rs.getString("name");
+                    String role = rs.getString("role");
+                    double winRate = rs.getDouble("winRate");
+                    c = new Champion(id, name, role, winRate);
                 }
-                if(ps != null){
-                    ps.close();
-                }
-                this.freeConnection(con);
             } catch (Exception e) {
                 throw new DAOexception("findChampionById() " + e.getMessage());
+            } finally {
+                try {
+                    if (rs != null) {
+                        rs.close();
+                    }
+                    if (ps != null) {
+                        ps.close();
+                    }
+                    this.freeConnection(con);
+                } catch (Exception e) {
+                    throw new DAOexception("findChampionById() " + e.getMessage());
+                }
             }
         }
         return c;
@@ -94,24 +113,26 @@ public class MySqlChampionDAO extends MySqlDAO implements IChampionDAO {
         PreparedStatement ps = null;
         int result = 0;
 
-        try {
-con = this.getConnection();
-            String query = "DELETE FROM champion WHERE id = ?";
-            ps = con.prepareStatement(query);
-            ps.setInt(1, id);
-            result = ps.executeUpdate();
-        } catch (Exception e) {
-            throw new DAOexception("deleteChampion() " + e.getMessage());
-        } finally {
+        if (ids.contains(id)){
             try {
-                if (ps != null) {
-                    ps.close();
-                }
-                this.freeConnection(con);
+                con = this.getConnection();
+                String query = "DELETE FROM champion WHERE id = ?";
+                ps = con.prepareStatement(query);
+                ps.setInt(1, id);
+                result = ps.executeUpdate();
             } catch (Exception e) {
                 throw new DAOexception("deleteChampion() " + e.getMessage());
-            }
+            } finally {
+                try {
+                    if (ps != null) {
+                        ps.close();
+                    }
+                    this.freeConnection(con);
+                } catch (Exception e) {
+                    throw new DAOexception("deleteChampion() " + e.getMessage());
+                }
 
+            }
         }
         return result == 1;
     }
@@ -144,5 +165,44 @@ con = this.getConnection();
             }
         }
         return result == 1;
+    }
+
+    @Override
+    public List<Champion> findChampionsByRole(String role) throws DAOexception {
+        Connection con = null;
+        PreparedStatement ps = null;
+        ResultSet rs = null;
+        List<Champion> champions = new ArrayList<>();
+
+        try {
+            con = this.getConnection();
+            String query = "SELECT * FROM champion WHERE role = ?";
+            ps = con.prepareStatement(query);
+            ps.setString(1, role);
+            rs = ps.executeQuery();
+
+            while(rs.next()){
+                int id = rs.getInt("id");
+                String name = rs.getString("name");
+                double winRate = rs.getDouble("win_rate");
+                Champion c = new Champion(id, name, role, winRate);
+                champions.add(c);
+            }
+        } catch (Exception e) {
+            throw new DAOexception("findChampionsByRole() " + e.getMessage());
+        } finally {
+            try{
+                if(rs != null){
+                    rs.close();
+                }
+                if(ps != null){
+                    ps.close();
+                }
+                this.freeConnection(con);
+            } catch (Exception e) {
+                throw new DAOexception("findChampionsByRole() " + e.getMessage());
+            }
+        }
+        return champions;
     }
 }
